+++ src/pages/programs/form/trackerProgram/fields/index.ts
@@ -1,0 +1,3 @@
+export { EnrollmentsLabelField } from './EnrollmentsLabelField'
+export { EventsLabelField } from './EventsLabelField'
+export { ProgramStagesLabelField } from './ProgramStagesLabelField'